rev version and fix tests
