+++ dwds/CHANGELOG.md
@@ -1,10 +1,13 @@
+## 3.1.2
+
+- Gracefully handle multiple injected clients on a single page.
+
 ## 3.1.1
 
 - Change the reported names for isolates to be more terse.
 - Implemented the 'PossibleBreakpoints' report kind for `getSourceReport()`.
 - Change the returned errors for the unimplemented `getClassList` and `reloadSources`
   methods to -32601 ('method does not exist / is not available').
-- Gracefully handle multiple injected clients on a single page.
 - Do not include native JavaScipt objects on stack returned from the debugger.
 
 ## 3.1.0
